chore: Moved buttons around
- Removed some redundant comments
