+++ mobile/src/components/new/Form/index.ts
@@ -1,8 +1,6 @@
-export { Button } from "./Button";
-export { IconButton } from "./IconButton";
+export { Button, IconButton, Ripple } from "./Button";
 export { NumericInput, TextInput } from "./Input";
 export { Legend, LegendItem } from "./Legend";
 export { ProgressBar } from "./ProgressBar";
-export { Ripple } from "./Ripple";
 export { Checkbox, Radio } from "./Selection";
 export { Switch } from "./Switch";
